requirements.txt
flask
python-dotenv
pandas
pillow
openai
weaviate-client